Added autoAliases to configuration
Valid options are 'allow' (default), 'warn' and 'restrict'

Now that we can import other chatito files it is very easy to use some
aliases from another file but forget to import that file which leads to
alias variants not being picked up.

We can use --autoAliases=warn to get warnings about auto-created aliases
to check if it's ok.
Or even use --autoAliases=restrict to disable the automatic aliases
creation. In this case we'll get an error for any alias that is not
defined